+++ apps/api/src/api/user/schemas/user.schema.ts
@@ -30,5 +30,10 @@ export const UserSchema = new Schema({
   selectedWorkspaceId: {
     type: Schema.Types.ObjectId,
     ref: "workspace"
+  },
+
+  isTrialAvailable: {
+    type: Boolean,
+    default: true
   }
 }, { versionKey: false })