+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ### v6.2.1
+* Removal of Hobby Plan & Introduction of Trial Plan
+* Dependency updates & bug fixes
 
 ### v6.2.0
 * Regular dependency updates & bug fixes